+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 * Transform scale tweening. ([32267c3](https://github.com/NazioLT/NTween/commit/32267c3e66aace669fde5b230537a643e3acbaae))
 * Custom callback tweening. ([3d5543e](https://github.com/NazioLT/NTween/commit/3d5543e3c161d68d7e0b3036afcd2df232a33ca5))
 * Color alpha tweening. ([abbce5f](https://github.com/NazioLT/NTween/commit/abbce5f85f83d1ce1e94ef54dac802148ed87005))
+* Adding waiting time before starting.